+++ frontend/app/components/Document/DocumentExplorer.tsx
@@ -19,12 +19,15 @@ import { MdOutlineRefresh } from "react-icons/md";
 import { MdContentPaste } from "react-icons/md";
 import { MdContentCopy } from "react-icons/md";
 import { TbVectorTriangle } from "react-icons/tb";
+import ContentView from "./ContentView";
 
 import { FaDatabase } from "react-icons/fa";
 import UserModalComponent from "../Navigation/UserModal";
 
 import { RAGConfig } from "../RAG/types";
 import ComponentStatus from "../Status/ComponentStatus";
+
+import { VerbaDocument, DocumentPayload } from "./types";
 
 interface DocumentExplorerProps {
   APIHost: string | null;
@@ -45,6 +48,54 @@ const DocumentExplorer: React.FC<DocumentExplorerProps> = ({
     "Content" | "Chunks" | "Metadata" | "Config" | "Vector Space" | "Graph"
   >("Content");
 
+  const [isFetching, setIsFetching] = useState(false);
+  const [document, setDocument] = useState<VerbaDocument | null>(null);
+
+  useEffect(() => {
+    if (selectedDocument) {
+      fetchSelectedDocument();
+    } else {
+      setDocument(null);
+    }
+  }, [selectedDocument]);
+
+  const fetchSelectedDocument = async () => {
+    try {
+      setIsFetching(true);
+
+      const response = await fetch(APIHost + "/api/get_document", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          uuid: selectedDocument,
+        }),
+      });
+
+      const data: DocumentPayload = await response.json();
+
+      if (data) {
+        if (data.error !== "") {
+          console.error(data.error);
+          setIsFetching(false);
+          setDocument(null);
+          setSelectedDocument(null);
+        } else {
+          setDocument(data.document);
+          setIsFetching(false);
+        }
+      }
+    } catch (error) {
+      console.error("Failed to fetch document:", error);
+      setIsFetching(false);
+    }
+  };
+
+  if (!selectedDocument) {
+    return <div></div>;
+  }
+
   return (
     <div className="flex flex-col gap-2 w-full">
       {/* Search Header */}
@@ -53,7 +104,7 @@ const DocumentExplorer: React.FC<DocumentExplorerProps> = ({
           <InfoComponent
             settingConfig={settingConfig}
             tooltip_text="Inspect your all information about your document, such as chunks, metadata and more."
-            display_text="Explorer"
+            display_text={document ? document.title : "Loading..."}
           />
         </div>
         <div className="flex gap-3 justify-end">
@@ -89,16 +140,6 @@ const DocumentExplorer: React.FC<DocumentExplorerProps> = ({
 
           <button
             onClick={() => {
-              setSelectedSetting("Graph");
-            }}
-            className={`flex ${selectedSetting === "Graph" ? "bg-primary-verba hover:bg-button-hover-verba" : "bg-button-verba hover:bg-button-hover-verba"} border-none btn text-text-verba gap-2`}
-          >
-            <SlGraph size={15} />
-            <p>Graph</p>
-          </button>
-
-          <button
-            onClick={() => {
               setSelectedDocument(null);
             }}
             className="flex btn btn-square border-none text-text-verba bg-button-verba hover:bg-warning-verba gap-2"
@@ -109,7 +150,18 @@ const DocumentExplorer: React.FC<DocumentExplorerProps> = ({
       </div>
 
       {/* Document List */}
-      <div className="bg-bg-alt-verba rounded-2xl flex flex-col p-6 items-center h-full w-full overflow-auto"></div>
+      <div className="bg-bg-alt-verba rounded-2xl flex flex-col p-6 h-full w-full overflow-y-auto overflow-x-hidden">
+        {isFetching && (
+          <div className="flex items-center justify-center text-text-alt-verba gap-2 h-full">
+            <span className="loading loading-spinner loading-sm"></span>
+            <p>Loading Document</p>
+          </div>
+        )}
+
+        {selectedSetting === "Content" && (
+          <ContentView document={document} settingConfig={settingConfig} />
+        )}
+      </div>
 
       {/* Import Footer */}
       <div className="bg-bg-alt-verba rounded-2xl flex gap-2 p-6 items-center justify-end h-min w-full">