+++ goldenverba/components/reader/UnstructuredAPI.py
@@ -8,8 +8,10 @@ from wasabi import msg
 
 from goldenverba.components.document import Document
 from goldenverba.components.interfaces import Reader
+from goldenverba.server.types import FileConfig
 
 from goldenverba.server.ImportLogger import LoggerManager
+from goldenverba.components.types import InputConfig
 
 
 class UnstructuredReader(Reader):
@@ -23,74 +25,77 @@ class UnstructuredReader(Reader):
         self.requires_env = ["UNSTRUCTURED_API_KEY"]
         self.name = "Unstructured"
         self.description = "Uses the Unstructured API to import multiple file types such as plain text and documents (.pdf, .csv). Requires an Unstructured API Key"
+        self.config = {
+            "API Key": InputConfig(
+                type="password", value="", description="You can set your Unstructured API Key here, it will overwrite the environment variable: UNSTRUCTURED_API_KEY", values=[]
+            ),
+            "Strategy": InputConfig(
+                type="dropdown",
+                value="auto",
+                description="Set the extraction strategy", values=["auto", "hi_res", "ocr_only", "fast"]
+            ),
+            "API URL": InputConfig(
+                type="text",
+                value="https://api.unstructured.io/general/v0/general",
+                description="Set the URL to the API", values=[]
+            ),
+        }
 
     async def load(
-        self, fileData: list, textValues: list[str], logger: LoggerManager
+        self, config: dict,  fileConfig: FileConfig
     ) -> list[Document]:
 
-        documents = []
+        document = None
 
-        url = os.environ.get(
-            "UNSTRUCTURED_API_URL", "https://api.unstructured.io/general/v0/general"
-        )
-        api_key = os.environ.get("UNSTRUCTURED_API_KEY", "")
+        if config["API Key"].value == "":
+            API_KEY = os.environ.get("UNSTRUCTURED_API_KEY")
+            if API_KEY is None:
+                raise Exception(f"No Unstructured API Key detected")
+        else:
+            API_KEY = config["API Key"].value
 
-        if api_key == "":
-            await logger.send_error(f"No Unstructed API Key detected")
-            msg.fail(f"No Unstructed API Key detected")
-            return []
+        API_URL = config["API URL"].value
+        STRATEGY = config["Strategy"].value
 
         headers = {
             "accept": "application/json",
-            "unstructured-api-key": api_key,
+            "unstructured-api-key": API_KEY,
         }
 
         data = {
-            "strategy": "auto",
+            "strategy": STRATEGY,
         }
 
-        for file in fileData:
-            msg.info(f"Loading in {file.filename}")
-            await logger.send_info(f"Importing {file.filename}")
+        msg.info(f"Loading in {fileConfig.filename}")
+        file_bytes = io.BytesIO(base64.b64decode(fileConfig.content))
+        file_data = {"files": (f"{fileConfig.filename}.{fileConfig.extension}", file_bytes)}
 
-            file_bytes = io.BytesIO(base64.b64decode(file.content))
-            file_data = {"files": (file.filename, file_bytes)}
+        try:
+            response = requests.post(
+                API_URL, headers=headers, data=data, files=file_data
+            )
+            json_response = response.json()
+            
+            if "detail" in json_response:
+                raise Exception(f"Failed to read {fileConfig.filename} : {json_response['detail']}")
 
-            try:
-                response = requests.post(
-                    url, headers=headers, data=data, files=file_data
-                )
-                json_response = response.json()
-                
-                if "detail" in json_response:
-                    msg.warn(
-                        f"Failed to load {file.filename} : {json_response['detail']}"
-                    )
-                    await logger.send_error(f"Failed to load {file.filename} : {json_response['detail']}")
-                    continue
+            full_content = ""
+            for chunk in json_response:
+                if "text" in chunk:
+                    text = str(chunk["text"])
+                    full_content += text + " "
 
-                full_content = ""
-                for chunk in json_response:
-                    if "text" in chunk:
-                        text = str(chunk["text"])
-                        full_content += text + " "
+            document = Document(
+                title=fileConfig.filename,
+                content=full_content,
+                extension=fileConfig.extension,
+                labels=fileConfig.labels,
+                source=fileConfig.source,
+                fileSize=fileConfig.file_size,
+                meta={}
+            )
 
-                if full_content == "":
-                    msg.warn(f"Empty Text for {file.filename}")
-                    await logger.send_warning(f"Empty Text for {file.filename}")
-                    continue
+            return document
 
-                document = Document(
-                    name=file.filename,
-                    text=full_content,
-                    type=self.config["document_type"].text,
-                    timestamp=str(datetime.now().strftime("%Y-%m-%d %H:%M:%S")),
-                    reader=self.name,
-                )
-                documents.append(document)
-
-            except Exception as e:
-                msg.warn(f"Failed to load {file.filename} : {str(e)}")
-                await logger.send_warning(f"Failed to load {file.filename} : {str(e)}")
-
-        return documents
+        except Exception as e:
+            raise Exception(f"Failed to load {fileConfig.filename} : {str(e)}")